requirements.txt
flask
google-genai
flask-cors
ollama